1. GoogleSignInHelper 2. refactoring
